requirements.txt
click
python-docx
elasticsearch
requests